requirements.txt
Django
djangorestframework
drf-yasg
numexpr
